renamed state  and some struct fields

+++ block/state.go
@@ -37,7 +37,7 @@ func (e *Executor) updateState(state types.State, block *types.Block, abciRespon
 		Version:         state.Version,
 		ChainID:         state.ChainID,
 		InitialHeight:   state.InitialHeight,
-		LastBlockHeight: int64(block.Header.Height),
+		LastBlockHeight: block.Header.Height,
 		LastBlockTime:   time.Unix(0, int64(block.Header.Time)),
 		LastBlockID: tmtypes.BlockID{
 			Hash: hash[:],
